interview_analyzer/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
openai-whisper>=20231117
pyAudioAnalysis>=0.3.14
spacy>=3.7.2
sentence-transformers>=2.2.2
ffmpeg-python>=0.2.0
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.27.0 